Split Astral Travel event handlers into 4 classes
